+++ examples/qpi_reconstruct.py
@@ -6,15 +6,16 @@
 import matplotlib.pyplot as plt
 
 from muscopy.backend_manager import BackendManager
-from muscopy.qpi import QPIParameters, make_disk, print_all_parameters, qpi
+from muscopy.dh import MuParameters, make_disk, print_all_parameters
+from muscopy.qpi import qpi
 
 # config
 SHOW_IMAGE = True
 
 # %%
-# set QPI parameters
+# set Microscopy parameters
 
-params = QPIParameters(
+params = MuParameters(
     na=0.8,
     wavelength_m=500e-9,
     img_size_px=512,